+++ instrumentation/net_http/lib/opentelemetry/instrumentation/net/http/patches/dup/instrumentation.rb
@@ -23,17 +23,20 @@ module OpenTelemetry
 
                 return super if untraced?
 
+                span_data = HttpHelper.span_attrs_for(req.method)
+
                 attributes = {
-                  OpenTelemetry::SemanticConventions::Trace::HTTP_METHOD => req.method,
+                  OpenTelemetry::SemanticConventions::Trace::HTTP_METHOD => span_data.normalized_method,
                   OpenTelemetry::SemanticConventions::Trace::HTTP_SCHEME => USE_SSL_TO_SCHEME[use_ssl?],
                   OpenTelemetry::SemanticConventions::Trace::HTTP_TARGET => req.path,
                   OpenTelemetry::SemanticConventions::Trace::NET_PEER_NAME => @address,
                   OpenTelemetry::SemanticConventions::Trace::NET_PEER_PORT => @port,
-                  'http.request.method' => req.method,
+                  'http.request.method' => span_data.normalized_method,
                   'url.scheme' => USE_SSL_TO_SCHEME[use_ssl?],
                   'server.address' => @address,
                   'server.port' => @port
                 }
+                attributes['http.request.method_original'] = span_data.original_method if span_data.original_method
                 path, query = split_path_and_query(req.path)
                 attributes['url.path'] = path
                 attributes['url.query'] = query if query
@@ -41,7 +44,7 @@ module OpenTelemetry
                 attributes.merge!(OpenTelemetry::Common::HTTP::ClientContext.attributes)
 
                 tracer.in_span(
-                  req.method,
+                  span_data.span_name,
                   attributes: attributes,
                   kind: :client
                 ) do |span|