+++ instrumentation/net_http/test/opentelemetry/instrumentation/net/http/old/instrumentation_test.rb
@@ -77,6 +77,29 @@ describe OpenTelemetry::Instrumentation::Net::HTTP::Instrumentation do
       )
     end
 
+    it 'after request with unknown http method' do
+      stub_request(:purge, 'http://example.com/purge').to_return(status: 200)
+      uri = URI('http://example.com/purge')
+      Net::HTTP.start(uri.host, uri.port) do |http|
+        http.request(Net::HTTP::Purge.new(uri))
+      end
+
+      _(exporter.finished_spans.size).must_equal 1
+      _(span.name).must_equal 'HTTP'
+      _(span.attributes['http.method']).must_equal '_OTHER'
+      _(span.attributes['http.scheme']).must_equal 'http'
+      _(span.attributes['http.status_code']).must_equal 200
+      _(span.attributes['http.target']).must_equal '/purge'
+      _(span.attributes['net.peer.name']).must_equal 'example.com'
+      _(span.attributes['net.peer.port']).must_equal 80
+      _(span.attributes['http.method.original']).must_be_nil
+      assert_requested(
+        :purge,
+        'http://example.com/purge',
+        headers: { 'Traceparent' => "00-#{span.hex_trace_id}-#{span.hex_span_id}-01" }
+      )
+    end
+
     it 'after request timeout' do
       expect do
         Net::HTTP.get(URI('https://example.com/timeout'))