+++ instrumentation/faraday/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History: opentelemetry-instrumentation-faraday
 
+### v0.30.1 / 2025-11-25
+
+* FIXED: Update support for unknown HTTP methods to match semantic conventions
+
 ### v0.30.0 / 2025-10-22
 
 * BREAKING CHANGE: Min Ruby Version 3.2